Split interactive submission choice into its own util function
