+++ grader/grader/commands/review.py
@@ -7,6 +7,7 @@ import re
 
 from grader.models import Grader
 from grader.utils.config import require_grader_config
+from grader.utils.interactive import submission_choice
 
 logger = logging.getLogger(__name__)
 
@@ -94,32 +95,6 @@ def review_loop(assignment, start_at):
             break
 
     print("Finished walking through all files.")
-
-
-def submission_choice(assignment, user_id, subs):
-    # FIXME: Combine code with inspect
-
-    # If they have multiple submissions, make them choose
-    if len(subs) > 1:
-        print("{0} submissions found for {1}, choose one:\n"
-              .format(len(subs), user_id))
-        i = 1
-        print("Index\tCreated")
-        for s in subs:
-            info = assignment.docker_cli.inspect_container(s.full_id)
-            print("{0}\t{1}".format(i, info['Created']))
-            i += 1
-        choice = -1
-        while choice < 0 or choice >= len(subs):
-            choice = input("Please enter your selection: ")
-            try:
-                choice = int(choice)-1
-            except TypeError:
-                pass
-
-        return subs[choice]
-    else:
-        return subs[0]
 
 
 def review_files(sub, editor):
